TODO -> running `copymove':  0. init.................. pass  1. begin................. FAIL (Could not create new collection `/davbox/ca/litmus/' for tests: 405 Method Not Allowed Server must allow `MKCOL /davbox/ca/litmus/' for tests to proceed) <- summary for `copymove': of 2 tests run: 1 passed, 1 failed. 50.0%

+++ src/main/java/cc/co/llabor/dav/cache/Cache4Dav.java
@@ -65,7 +65,8 @@ public  class Cache4Dav extends AbstractTransactionalDaver implements IWebdavSto
  
 			// delete file
 			try{
-				MemoryFileItem toDel = memFS.get(uri); 
+				MemoryFileItem toDel = memFS.get(uri);
+				if (toDel==null)return;
 				System_out_println("<delete>\n\t<file name=\'"+toDel.getName()+"\'/>... ");
 				toDel.delete();
 				Object retval = memFS.delete(toDel);
